put all methods together in a single file compute_ttr.py
