+++ UncertainSCI/ex_pws.py
@@ -1,16 +1,14 @@
 import numpy as np
 
-from UncertainSCI.mthd_hankel_det import hankel_det
+from UncertainSCI.compute_ttr import predict_correct_bounded, stieltjes_bounded, \
+        aPC_bounded, hankel_det, mod_cheb
 
-from UncertainSCI.mthd_mod_cheb import mod_cheb
-from UncertainSCI.mthd_mod_correct import gq_modification_composite
+from UncertainSCI.utils.compute_mom import compute_mom_bounded
+
+from UncertainSCI.utils.compute_subintervals import compute_subintervals
+from UncertainSCI.utils.quad import gq_modification_composite
 from UncertainSCI.families import JacobiPolynomials
 
-from UncertainSCI.mthd_mod_correct import compute_subintervals, compute_ttr_bounded
-
-from UncertainSCI.mthd_stieltjes import stieltjes_bounded
-
-from UncertainSCI.mthd_aPC import aPC_bounded, compute_mom_bounded
 
 import scipy.integrate as integrate
 import scipy.special as sp
@@ -21,11 +19,11 @@ from tqdm import tqdm
 """
 We use five methods
 
-1. hankel_det (Hankel determinant)
-2. mod_cheb (modified Chebyshev)
-3. mod_correct (modified correct)
-4. stieltjes (Stieltjes)
-5. aPC (arbitrary polynomial chaos expansion)
+1. predict_correct (Predict-Correct)
+2. stieltjes (Stieltjes)
+3. aPC (Arbitrary Polynomial Chaos Expansion)
+4. hankel_det (Hankel Determinant)
+5. mod_cheb (Modified Chebyshev)
 
 to compute the recurrence coefficients for the piecewise weight function.
 """
@@ -128,17 +126,17 @@ singularity_list = [ [-1, 0, q],
 
 N_array = [10, 20, 40, 80]
 
-t_hankel_det = np.zeros(len(N_array))
-t_mod_cheb = np.zeros(len(N_array))
-t_mod_correct = np.zeros(len(N_array))
+t_predict_correct = np.zeros(len(N_array))
 t_stieltjes = np.zeros(len(N_array))
 t_aPC = np.zeros(len(N_array))
+t_hankel_det = np.zeros(len(N_array))
+t_mod_cheb = np.zeros(len(N_array))
 
-l2_hankel_det = np.zeros(len(N_array))
-l2_mod_cheb = np.zeros(len(N_array))
-l2_mod_correct = np.zeros(len(N_array))
+l2_predict_correct = np.zeros(len(N_array))
 l2_stieltjes = np.zeros(len(N_array))
 l2_aPC = np.zeros(len(N_array))
+l2_hankel_det = np.zeros(len(N_array))
+l2_mod_cheb = np.zeros(len(N_array))
 
 iter_n = np.arange(10)
 for k in tqdm(iter_n):
@@ -147,15 +145,37 @@ for k in tqdm(iter_n):
         
         ab = ab_pws4(N)[:N]
 
-        # Hankel determinant
-        mom = compute_mom_bounded(a, b, weight, N+1, singularity_list)
+        m = compute_mom_bounded(a, b, weight, N, singularity_list)
+
+        # Predict-Correct
         start = time.time()
-        ab_hankel_det = hankel_det(N = N, mom = mom)
+        ab_predict_correct = predict_correct_bounded(a, b, weight, N, singularity_list)
+        end = time.time()
+        t_predict_correct[ind] += (end - start) / len(iter_n)
+        l2_predict_correct[ind] += np.linalg.norm(ab - ab_predict_correct, None) / len(iter_n)
+
+        # Stieltjes
+        start = time.time()
+        ab_stieltjes = stieltjes_bounded(a, b, weight, N, singularity_list)
+        end = time.time()
+        t_stieltjes[ind] += (end - start) / len(iter_n)
+        l2_stieltjes[ind] += np.linalg.norm(ab - ab_stieltjes, None) / len(iter_n)
+
+        # Arbitrary Polynomial Chaos Expansion
+        start = time.time()
+        ab_aPC = aPC_bounded(a, b, weight, N, singularity_list, m)
+        end = time.time()
+        t_aPC[ind] += (end - start) / len(iter_n)
+        l2_aPC[ind] += np.linalg.norm(ab - ab_aPC, None) / len(iter_n)
+
+        # Hankel Determinant
+        start = time.time()
+        ab_hankel_det = hankel_det(N, m)
         end = time.time()
         t_hankel_det[ind] += (end - start) / len(iter_n)
         l2_hankel_det[ind] += np.linalg.norm(ab - ab_hankel_det, None) / len(iter_n)
 
-        # modified Chebyshev
+        # Modified Chebyshev
         J = JacobiPolynomials(probability_measure=False)
         peval = lambda x, n: J.eval(x, n)
         subintervals = compute_subintervals(a, b, singularity_list)
@@ -169,27 +189,6 @@ for k in tqdm(iter_n):
         t_mod_cheb[ind] += (end - start) / len(iter_n)
         l2_mod_cheb[ind] += np.linalg.norm(ab - ab_mod_cheb, None) / len(iter_n)
 
-        # modified correct
-        start = time.time()
-        ab_mod_correct = compute_ttr_bounded(a, b, weight, N, singularity_list)
-        end = time.time()
-        t_mod_correct[ind] += (end - start) / len(iter_n)
-        l2_mod_correct[ind] += np.linalg.norm(ab - ab_mod_correct, None) / len(iter_n)
-
-        # Stieltjes
-        start = time.time()
-        ab_stieltjes = stieltjes_bounded(a, b, weight, N, singularity_list)
-        end = time.time()
-        t_stieltjes[ind] += (end - start) / len(iter_n)
-        l2_stieltjes[ind] += np.linalg.norm(ab - ab_stieltjes, None) / len(iter_n)
-
-        # arbitrary polynomial chaos expansion
-        start = time.time()
-        ab_aPC = aPC_bounded(a, b, weight, N, singularity_list)
-        end = time.time()
-        t_aPC[ind] += (end - start) / len(iter_n)
-        l2_aPC[ind] += np.linalg.norm(ab - ab_aPC, None) / len(iter_n)
-
 
 """
 N_array = [10, 20, 40, 80] with tol = 1e-12
@@ -198,29 +197,38 @@ case pws1 (gm = 1, p = q = -1/2)
 
 --- l2 error ---
 
-l2_hankel_det
-array([9.26628527e-10, 6.04630201e-02,            nan,            nan])
-l2_mod_cheb
-array([1.53509567e-15, 2.33645016e-15, 1.00191298e+00,            nan])
-l2_mod_correct
+l2_predict_correct
 array([5.00951261e-15, 9.07801721e-15, 1.80114070e-14, 5.14165169e-14])
+
 l2_stieltjes
 array([8.18521290e-15, 4.73425186e-14, 2.85017480e-13, 3.99304271e-13])
+
 l2_aPC
 array([2.70808490e-11, 1.57229820e-02, 3.50690996e+00, 1.67875477e+01])
 
+l2_hankel_det
+array([9.26628527e-10, 6.04630201e-02,            nan,            nan])
+
+l2_mod_cheb
+array([1.53509567e-15, 2.33645016e-15, 1.00191298e+00,            nan])
+
+
 --- elapsed time ---
 
-t_hankel_det
-array([0.00119677, 0.00298259, 0.00981224, 0.0349107 ])
-t_mod_cheb
-array([0.00038941, 0.00160246, 0.00649621, 0.02559817])
-t_mod_correct
-array([0.04332623, 0.10618875, 0.30466142, 0.97284315])
+t_predict_correct
+array([0.04977272, 0.10818396, 0.29909995, 0.96135225])
+
 t_stieltjes
-array([0.04452567, 0.11023769, 0.29800267, 0.94145155])
+array([0.04228184, 0.10623031, 0.29446549, 0.96905365])
+
 t_aPC
-array([0.09312241, 0.23187673, 0.68196747, 2.39598153])
+array([0.05775449, 0.14411304, 0.45337014, 1.5779732 ])
+
+t_hankel_det
+array([0.00094748, 0.00260286, 0.00890198, 0.03357236])
+
+t_mod_cheb
+array([0.00043089, 0.00160668, 0.00627456, 0.02566364])
 
 
 
@@ -228,29 +236,38 @@ case pws2 (gm = -1, p = q = -1/2)
 
 --- l2 error ---
 
-l2_hankel_det
-array([2.22385227e-10, 1.32503881e-02,            nan,            nan])
-l2_mod_cheb
-array([4.76049343e-11, 6.81367106e-11, 9.31905279e-01,            nan])
-l2_mod_correct
+l2_predict_correct
 array([4.71520472e-11, 4.71520480e-11, 4.71520506e-11, 4.71520742e-11])
+
 l2_stieltjes
 array([4.71518630e-11, 4.71519331e-11, 4.71531130e-11, 4.71550346e-11])
+
 l2_aPC
 array([1.40680558e-11, 3.49835593e-03, 5.03089057e+00, 1.48071565e+01])
 
+l2_hankel_det
+array([2.22385227e-10, 1.32503881e-02,            nan,            nan])
+
+l2_mod_cheb
+array([4.76049343e-11, 6.81367106e-11, 9.31905279e-01,            nan])
+
+
 --- elapsed time ---
 
-t_hankel_det
-array([0.00129924, 0.00273745, 0.00963595, 0.03425987])
-t_mod_cheb
-array([0.00041225, 0.00151677, 0.00632379, 0.02551525])
-t_mod_correct
-array([0.04608409, 0.10838559, 0.29906871, 0.94933126])
+t_predict_correct
+array([0.0523699 , 0.11145608, 0.30043929, 0.94330449])
+
 t_stieltjes
-array([0.04356062, 0.1036242 , 0.29459953, 0.93659427])
+array([0.04802203, 0.1058763 , 0.30032334, 0.93709919])
+
 t_aPC
-array([0.09356785, 0.2234241 , 0.65123837, 2.31522679])
+array([0.06006751, 0.15445635, 0.45450728, 1.51733439])
+
+t_hankel_det
+array([0.00093138, 0.00262151, 0.00880477, 0.03218133])
+
+t_mod_cheb
+array([0.00039856, 0.00171826, 0.0062541 , 0.02528358])
 
 
 
@@ -258,29 +275,39 @@ case pws3 (gm = 1, p = q = 1/2)
 
 --- l2 error ---
 
-l2_hankel_det
-array([5.70934219e-10, 6.47300662e-03,            nan,            nan])
-l2_mod_cheb
-array([1.65136200e-15, 7.14539839e-12,            nan,            nan])
-l2_mod_correct
+l2_predict_correct
 array([2.60321138e-15, 5.95361983e-15, 1.58834452e-14, 3.90074024e-14])
+
 l2_stieltjes
 array([5.52914474e-15, 1.83470863e-14, 1.29512727e-13, 3.22031853e-13])
+
 l2_aPC
 array([7.08100353e-11, 3.48873407e-04, 4.76686828e+00, 2.40114640e+01])
 
+l2_hankel_det
+array([5.70934219e-10, 6.47300662e-03,            nan,            nan])
+
+l2_mod_cheb
+array([1.65136200e-15, 7.14539839e-12,            nan,            nan])
+
+
 --- elapsed time ---
 
-t_hankel_det
-array([0.00120425, 0.00305567, 0.01018758, 0.03417921])
-t_mod_cheb
-array([0.00043008, 0.00160787, 0.00655656, 0.02591875])
-t_mod_correct
-array([0.04156477, 0.10791004, 0.30081213, 0.97484508])
+t_predict_correct
+array([0.04465756, 0.10191462, 0.29291177, 0.94399831])
+
 t_stieltjes
-array([0.04191537, 0.10519171, 0.29755795, 0.99244423])
+array([0.04031701, 0.10070646, 0.28847523, 0.95105956])
+
 t_aPC
-array([0.08718987, 0.21545782, 0.71292422, 2.38806105])
+array([0.05571589, 0.13915374, 0.43480496, 1.5230818 ])
+
+t_hankel_det
+array([0.00099432, 0.00271358, 0.00898612, 0.03350439])
+
+t_mod_cheb
+array([0.00040579, 0.00152814, 0.00634031, 0.02579982])
+
 
 
 
@@ -288,29 +315,38 @@ case pws4 (gm = -1, p = q = 1/2)
 
 --- l2 error ---
 
-l2_hankel_det
-array([1.61824859e-10, 1.26143221e-03,            nan,            nan])
-l2_mod_cheb
-array([1.03122952e-11, 7.35418481e-11, 1.01409235e+00,            nan])
-l2_mod_correct
+l2_predict_correct
 array([3.56442518e-12, 3.56442872e-12, 3.56446526e-12, 3.56464316e-12])
+
 l2_stieltjes
 array([3.56095591e-12, 3.56099734e-12, 3.56222887e-12, 3.56641569e-12])
+
 l2_aPC
 array([2.33008463e-11, 3.39461459e-05, 4.36976935e+00, 1.82287098e+01])
 
+l2_hankel_det
+array([1.61824859e-10, 1.26143221e-03,            nan,            nan])
+
+l2_mod_cheb
+array([1.03122952e-11, 7.35418481e-11, 1.01409235e+00,            nan])
+
+
 --- elapsed time ---
 
-t_hankel_det
-array([0.00128016, 0.0026835 , 0.01041219, 0.03428562])
-t_mod_cheb
-array([0.00039186, 0.00152152, 0.00645249, 0.02545857])
-t_mod_correct
-array([0.04417112, 0.11201956, 0.30561185, 0.95412657])
+t_predict_correct
+array([0.04811735, 0.10311818, 0.29336064, 0.94649303])
+
 t_stieltjes
-array([0.04193373, 0.11085222, 0.29145977, 0.9568953 ])
+array([0.042046  , 0.10245237, 0.29115927, 0.94028342])
+
 t_aPC
-array([0.09121854, 0.23482001, 0.65988688, 2.37069175])
+array([0.06035998, 0.1419471 , 0.44453773, 1.52734239])
+
+t_hankel_det
+array([0.00095551, 0.00270836, 0.00892756, 0.03262351])
+
+t_mod_cheb
+array([0.00039389, 0.00150034, 0.00632353, 0.02536793])
 
 """
